# docs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Ekklesia Voting Backend API — v1
  description: |
    Hydra-backed live surface for the Ekklesia voting platform. v1 is paired
    with the v0 archival surface (see `openapi.yaml`) — the two coexist.

    ## What lives here

    - **Unified ballot listing** across legacy + Hydra sources.
    - **Broker endpoints** for Hydra votes: draft the signing payload,
      collect signatures (key-based or m-of-n multisig), submit to Hydra,
      surface confirmation artifacts.
    - **Admin lifecycle**: `/prepare`, `/start`, `/finalize`, the stepped
      close sequence (`/settle/burn` → `/settle/finalize` → `/settle/close`),
      plus inspection helpers (`/count`, `/queue/*`, `/head-info`,
      `/results/recover`) and `/voting-power` + `/certify` — role-gated.
    - **Public integrator surface** under `/public/*` — API-key gated,
      rate-limited, read-only.

    ## Authentication

    - `verifyToken` (JWT cookie or Authorization header) for voter session
      endpoints.
    - Same JWT + admin allowlist/claim for `/admin/*`.
    - API key (Bearer or x-api-key) for `/public/*`.
  version: 1.0.0
  contact:
    name: Adam Dean, Mad Orkestra
    license: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://docs.ekklesia.vote/api
    description: Documentation reference host
...
paths:
...
  /ballots:
    get:
      tags: [Unified Ballots]
      summary: List ballots across all sources
      description: |
        `Cache-Control: public, max-age=60`. Ballot definitions are largely
        static once published; poll **/results** separately for live tallies.
      parameters:
        - in: query
          name: voterType
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [upcoming, live, closed] }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - in: query
          name: source
          description: Restrict to a single adapter (`legacy` or `hydra`).
          schema: { type: string, enum: [legacy, hydra] }
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/UnifiedBallot' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
  /ballots/{id}:
    get:
      tags: [Unified Ballots]
      summary: Ballot detail (enriched from Hydra when source === "hydra")
      description: |
        Cache-Control varies by status: `closed` → `max-age=3600`, `live` →
        `max-age=120`, `upcoming` → `max-age=30`. Results are **not** embedded —
        call `/results/ballot/{id}` separately.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Ballot
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/UnifiedBallot' }
        '404': { description: Not found }
  /ballots/{id}/questions/{qid}/content:
    get:
      tags: [Unified Ballots, Audit]
      summary: Per-proposal canonical content bytes (audit chain-of-custody)
      description: |
        Returns the byte-identical canonical JSON that `Proposal.contentHash`
        was computed over. Auditors re-hash these bytes with `blake2b_256` to
        verify the proposal hasn't drifted since ballot-prepare time.

        Public, unauthenticated by design — long-term auditability requires
        no gate. The chain of custody is:

          on-chain (600) datum → ekklesia.merkleRoot
            → IPFS-pinned ballot JSON (covers BallotQuestion.contentHash[n])
              → per-proposal content blob (this endpoint).

        Cache-Control mirrors `/ballots/{id}` and varies by ballot status.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: qid
          required: true
          description: Proposal id (must belong to the ballot).
          schema: { type: string }
      responses:
        '200':
          description: Canonical content bytes
          content:
            application/json:
              schema:
                type: object
                description: Canonical proposal content blob; byte stream is the authoritative artifact.
        '404': { description: Ballot or proposal not found }
  /ballots/{id}/archive:
    get:
      tags: [Unified Ballots, Audit]
      summary: Full audit bundle (ballot + proposals + manifest + README)
      description: |
        One JSON file carrying every voter-facing committed field, a
        per-file hash MANIFEST (blake2b_256 + sha256), and a README with
        the verification recipe. Designed to be saved to disk and re-pinned
        to IPFS by anyone who cares about long-term audit. Public,
        unauthenticated.

        The response is served with `Content-Disposition: attachment` so
        browsers download instead of preview.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Audit bundle
          content:
            application/json:
              schema:
                type: object
                required: [schemaVersion, generatedAt, ballot, proposals, manifest, readme]
                properties:
                  schemaVersion: { type: string, example: '1' }
                  generatedAt: { type: string, format: date-time }
                  ballot:
                    type: object
                    description: Voter-facing ballot summary (id, title, voter window, hydra anchors).
                  proposals:
                    type: object
                    description: Map of proposalId → canonical content blob.
                    additionalProperties: { type: object }
                  manifest:
                    type: array
                    description: Per-file hash entries — `ballot.json` plus `proposals/<id>.json`.
                    items:
                      type: object
                      required: [path, bytes, blake2b_256, sha256]
                      properties:
                        path: { type: string }
                        bytes: { type: integer }
                        blake2b_256: { type: string }
                        sha256: { type: string }
                  readme:
                    { type: string, description: Verification recipe + chain-of-custody notes. }
        '404': { description: Ballot not found }
  /ballots/{id}/certified:
    get:
      tags: [Unified Ballots, Results]
      summary: Authority certification state + version history
      description: |
        Surfaces the currently-active `CertifiedSnapshot` (if any) plus the
        full version history so the frontend can label results as
        "Certified at version N by <authority>" vs. "Provisional
        (Hydra-final, not yet certified)".

        - `certified: true`  → an authority snapshot is active; full
          shape includes per-proposal tallies and snapshot metadata.
        - `certified: false` → no active certification. `narrative` may
          still be set if a narrative-only publication exists.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Certification state
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: 'success' }
                  data:
                    oneOf:
                      - type: object
                        required: [certified, history]
                        properties:
                          certified: { type: boolean, enum: [false] }
                          narrative:
                            nullable: true
                            type: object
                            properties:
                              url: { type: string }
                              label: { type: string }
                          history:
                            type: array
                            items: { $ref: '#/components/schemas/CertifiedHistoryEntry' }
                      - type: object
                        required: [certified, version, certifiedAt, history]
                        properties:
                          certified: { type: boolean, enum: [true] }
                          version: { type: integer }
                          certifiedAt: { type: string, format: date-time }
                          snapshotUrl: { type: string, nullable: true }
                          snapshotHash: { type: string, nullable: true }
                          snapshotEpoch: { type: integer, nullable: true }
                          narrative:
                            nullable: true
                            type: object
                            properties:
                              url: { type: string }
                              label: { type: string }
                          perProposal:
                            type: array
                            items:
                              type: object
                              properties:
                                proposalId: { type: string }
                                results:
                                  type: array
                                  items: { type: object }
                                resultsByGroup: { type: object }
                          history:
                            type: array
                            items: { $ref: '#/components/schemas/CertifiedHistoryEntry' }
        '404': { description: Ballot not found }
  /results/ballot/{ballotId}:
    get:
      tags: [Results]
      summary: All Result rows for a ballot — anonymous, short-cached
      description: |
        Served with `Cache-Control: public, max-age=30` so clients can poll
        cheaply while the cron updates provisional tallies every ~10 min.
        Final tallies land the moment `/admin/ballots/{id}/finalize` completes.
      parameters:
        - in: path
          name: ballotId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Array of Result rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Result' }
...
  /proposals/ballot/{ballotId}:
    get:
      tags: [Proposals]
      summary: List proposals for a ballot, with facet-driven sort + filter
      description: |
        Sort and filter keys must be declared on `Ballot.facets[]`.
        Multi-value enum filters take CSV (e.g. `?filter[category]=a,b`);
        OR semantics. Unknown keys and attempts to sort on a
        non-sortable facet return 400. Facets are frozen once the
        ballot goes live, so the filter UI can safely cache the
        facet list for the ballot's lifetime.
      parameters:
        - in: path
          name: ballotId
          required: true
          schema: { type: string }
        - in: query
          name: sort
          description: 'Facet key to sort by. Must have `sortable: true`.'
          schema: { type: string }
        - in: query
          name: dir
          schema: { type: string, enum: [asc, desc] }
        - in: query
          name: filter
          style: deepObject
          explode: true
          description: |
            Object of `filter[<facetKey>]=<csv>` filters. CSV split on
            `,`; OR within a single facet; AND across facets.
          schema:
            type: object
            additionalProperties: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Paginated proposals
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: 'success' }
                  data: { type: array, items: { type: object } }
                  pagination:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                  applied:
                    type: object
                    properties:
                      filters:
                        type: object
                        additionalProperties:
                          type: array
                          items: { type: string }
                      sort:
                        type: object
                        properties:
                          key: { type: string }
                          direction: { type: string, enum: [asc, desc] }
                          source: { type: string, enum: [default, fallback] }
        '400':
          description: Unknown facet key, unsortable sort target, or bad filter value
        '404': { description: Ballot not found }
...
  /public/ballots:
    get:
      tags: [Public]
      summary: Public unified ballot listing (API-key gated, rate-limited)
      security: [{ apiKeyHeader: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: source
          schema: { type: string, enum: [legacy, hydra] }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/UnifiedBallot' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { description: Missing or invalid API key }
        '403': { description: Scope mismatch }
        '429': { description: Rate limit exceeded }
  /public/ballots/{id}:
    get:
      tags: [Public]
      summary: Public ballot detail
      security: [{ apiKeyHeader: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
  /public/results/ballot/{ballotId}:
    get:
      tags: [Public]
      summary: All Result rows for a ballot (provisional + final)
      security: [{ apiKeyHeader: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: ballotId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Array of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Result' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    UnifiedBallot:
      type: object
      properties:
        id: { type: string }
        source: { type: string, enum: [legacy, hydra] }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [upcoming, live, closed] }
        voterType: { type: string }
        voterGroups:
          type: array
          description: >
            Per-group eligibility + power-source declaration. Each entry
            pairs a role key (drep/pool/stake) with a Hydra
            RoleWeighting value (CredentialBased, StakeBased, or
            PledgeBased). Translated directly into Hydra's roleWeighting
            object at /prepare. Valid combinations:
            drep→CredentialBased|StakeBased;
            pool→CredentialBased|StakeBased|PledgeBased;
            stake→StakeBased.
          items:
            type: object
            required: [group, powerSource]
            properties:
              group: { type: string, enum: [drep, pool, stake] }
              powerSource: { type: string, enum: [CredentialBased, StakeBased, PledgeBased] }
        voterDescription: { type: string }
        voteWeighted: { type: boolean }
        votePeriodStart: { type: string, format: date-time }
        votePeriodEnd: { type: string, format: date-time }
        voteFilters: { type: boolean }
        ipfsHash: { type: string, nullable: true }
        proposalCount: { type: integer, nullable: true }
        singleProposal: { type: string, nullable: true }
        hydra:
          type: object
          nullable: true
          properties:
            endpoint: { type: string, nullable: true }
            headId: { type: string, nullable: true }
            ballotCid: { type: string, nullable: true }
            instancePolicyId: { type: string, nullable: true }
            headInfo: { type: object, nullable: true }
            ballot: { type: object, nullable: true }
        provisionalResultsEnabled: { type: boolean }
        proposalSource:
          nullable: true
          description: |
            Present when the ballot was imported from an upstream
            proposals module. Null for scaffold-created or legacy
            ballots.
          type: object
          properties:
            moduleId: { type: string }
            moduleUrl: { type: string, nullable: true }
            externalBallotId: { type: string }
            version: { type: string, nullable: true }
            importedAt: { type: string, format: date-time }
            importMethod: { type: string, enum: [push, upload] }
            importedBy: { type: string }
        facets:
          type: array
          description: Sort/filter dimensions declared for this ballot.
          items: { $ref: '#/components/schemas/FacetDefinition' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    FacetDefinition:
      type: object
      required: [key, label, type]
      description: |
        Declares a sort/filter dimension for this ballot's proposals.
        Frozen at `status === "live"`. CSV is the wire format for
        multi-value enums (option names must not contain `,`).
      properties:
        key:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
        label: { type: string, maxLength: 120 }
        type:
          type: string
          enum: [enum, number, string, boolean, date]
        multi:
          type: boolean
          default: false
          description: |
            Only meaningful for `type:"enum"`. Implies `sortable:false`.
        options:
          type: array
          description: Required when `type:"enum"`. Items must not contain `,`.
          items: { type: string }
        unit: { type: string, nullable: true }
        sortable: { type: boolean, default: false }
        filterable: { type: boolean, default: true }
        defaultSort:
          type: string
          enum: [asc, desc]
          nullable: true
          description: At most one facet may declare this.
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
...
    CertifiedHistoryEntry:
      type: object
      description: One row of a ballot's certification history (newest-first ordering).
      properties:
        version: { type: integer }
        submittedAt: { type: string, format: date-time }
        submittedBy: { type: string }
        source: { type: string, description: 'Origin of the certification (e.g. "api", "chain").' }
        chainTxHash: { type: string, nullable: true }
        snapshotUrl: { type: string, nullable: true }
        snapshotHash: { type: string, nullable: true }
        narrativeOnly: { type: boolean }
        narrative:
          nullable: true
          type: object
          properties:
            url: { type: string }
            label: { type: string }
...